offer accept and reject updating ui
