+++ src/Pages/MyApplications/styles.scss
@@ -38,7 +38,7 @@
           background: rgb(227, 126, 18);
         }
         &.closedIcon {
-          background: gray;
+          background: rgb(12, 3, 87);
         }
       }
       .info {
